compose: Remove lock-padding class.
This was previously needed because of the way the icon was
added to the stream input field. Now that the icon is inside
the dropdown menu option, it no longer needs to be 0 width,
and can handle the padding between it and the sream name.

Note this applies to both the lock and globe icons, which
were both using "lock-padding".